Added SQL TextArea
Added the SQL TextArea and validation but I broke the formatting of the form. I need to figure out how to show the expression builder nicely without items overflowing. I also need to look into using the jimu SQLExpressionBuilder. Right now, it will throw an error when adding a clause.